% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_lr_gmt}
\alias{read_lr_gmt}
\title{Read ligand-receptor pathway sets in GMT format}
\usage{
read_lr_gmt(file)
}
\arguments{
\item{file}{Path to a tab-delimited GMT file.}
}
\value{
A named list of interaction identifiers. The second GMT field is ignored.
}
\description{
Read ligand-receptor pathway sets in GMT format
}
